extract search bar out to reusable component
